refactor-settings * Rename settings/index.tsx to settings/useFirebaseSettings.ts
