+++ src/containers/Admin/EditTab/EditTab.tsx
@@ -17,7 +17,7 @@ import { Switch, TextField } from '@entur/form'
 import { Tooltip } from '@entur/tooltip'
 import { ValidationInfoIcon } from '@entur/icons'
 import { Button } from '@entur/button'
-import { Mode } from '../../../settings'
+import { Mode } from '../../../settings/useFirebaseSettings'
 import { useSettings } from '../../../settings/SettingsProvider'
 import { isMobileWeb, getTranslation } from '../../../utils/utils'
 import { DEFAULT_DISTANCE, DEFAULT_ZOOM } from '../../../constants'
