+++ src/containers/MyBoards/OwnedBoards/BoardCard/BoardCard.tsx
@@ -5,7 +5,7 @@ import { Heading3 } from '@entur/typography'
 import { LinkIcon, ClockIcon } from '@entur/icons'
 import { ThemeDashboardPreview } from '../../../../assets/icons/ThemeDashboardPreview'
 import { persistSingleField } from '../../../../settings/FirestoreStorage'
-import { Settings } from '../../../../settings'
+import { Settings } from '../../../../settings/useFirebaseSettings'
 import { createTimeString } from '../../../../utils/time'
 import { BoardOverflowMenu } from './OverflowMenu/BoardOverflowMenu'
 import './BoardCard.scss'
