+++ src/services/firebase.ts
@@ -26,7 +26,7 @@ import { getDownloadURL, ref, uploadBytesResumable } from 'firebase/storage'
 import type { User } from 'firebase/auth'
 import { signInWithCustomToken } from 'firebase/auth'
 import { auth, db, functions, storage } from '../firebase-init'
-import { Settings } from '../settings'
+import { Settings } from '../settings/useFirebaseSettings'
 import { FieldTypes } from '../settings/FirestoreStorage'
 import { BoardOwnersData } from '../types'
